backend/requirements.txt
Django>=5.2
psycopg2-binary
djangorestframework
django-cors-headers
djangorestframework-simplejwt
